+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/guide/ChestSlimefunGuide.java
@@ -353,7 +353,7 @@ public class ChestSlimefunGuide implements SlimefunGuideImplementation {
             return;
         }
 
-        Recipe[] recipes = SlimefunPlugin.getMinecraftRecipes().getRecipesFor(item).toArray(new Recipe[0]);
+        Recipe[] recipes = SlimefunPlugin.getMinecraftRecipes().getRecipesFor(item);
 
         if (recipes.length == 0) {
             return;
@@ -375,7 +375,7 @@ public class ChestSlimefunGuide implements SlimefunGuideImplementation {
         if (optional.isPresent()) {
             MinecraftRecipe<?> mcRecipe = optional.get();
 
-            RecipeChoice[] choices = SlimefunPlugin.getMinecraftRecipes().getRecipeInput(recipe);
+            RecipeChoice[] choices = SlimefunPlugin.getMinecraftRecipes().getRecipeShape(recipe);
 
             if (choices.length == 1 && choices[0] instanceof MaterialChoice) {
                 recipeItems[4] = new ItemStack(((MaterialChoice) choices[0]).getChoices().get(0));