+++ src/main/java/me/mrCookieSlime/Slimefun/SlimefunPlugin.java
@@ -14,7 +14,6 @@ import org.bukkit.plugin.java.JavaPlugin;
 
 import io.github.thebusybiscuit.cscorelib2.config.Config;
 import io.github.thebusybiscuit.cscorelib2.protection.ProtectionManager;
-import io.github.thebusybiscuit.cscorelib2.recipes.RecipeSnapshot;
 import io.github.thebusybiscuit.cscorelib2.reflection.ReflectionUtils;
 import io.github.thebusybiscuit.slimefun4.api.SlimefunAddon;
 import io.github.thebusybiscuit.slimefun4.api.gps.GPSNetwork;
@@ -27,6 +26,7 @@ import io.github.thebusybiscuit.slimefun4.core.services.BlockDataService;
 import io.github.thebusybiscuit.slimefun4.core.services.CustomItemDataService;
 import io.github.thebusybiscuit.slimefun4.core.services.CustomTextureService;
 import io.github.thebusybiscuit.slimefun4.core.services.LocalizationService;
+import io.github.thebusybiscuit.slimefun4.core.services.MinecraftRecipeService;
 import io.github.thebusybiscuit.slimefun4.core.services.PermissionsService;
 import io.github.thebusybiscuit.slimefun4.core.services.UpdaterService;
 import io.github.thebusybiscuit.slimefun4.core.services.github.GitHubService;
@@ -97,15 +97,16 @@ public final class SlimefunPlugin extends JavaPlugin implements SlimefunAddon {
 
     // Services - Systems that fulfill certain tasks, treat them as a black box
     private final CustomItemDataService itemDataService = new CustomItemDataService(this, "slimefun_item");
-    private final CustomTextureService textureService = new CustomTextureService(this);
     private final BlockDataService blockDataService = new BlockDataService(this, "slimefun_block");
+    private final CustomTextureService textureService = new CustomTextureService(this);
     private final GitHubService gitHubService = new GitHubService("TheBusyBiscuit/Slimefun4");
     private final UpdaterService updaterService = new UpdaterService(this, getFile());
+    private final MetricsService metricsService = new MetricsService(this);
     private final AutoSavingService autoSavingService = new AutoSavingService();
     private final BackupService backupService = new BackupService();
-    private final MetricsService metricsService = new MetricsService(this);
     private final PermissionsService permissionsService = new PermissionsService(this);
     private final ThirdPartyPluginService thirdPartySupportService = new ThirdPartyPluginService(this);
+    private final MinecraftRecipeService recipeService = new MinecraftRecipeService(this);
     private LocalizationService local;
 
     private GPSNetwork gpsNetwork;
@@ -114,7 +115,6 @@ public final class SlimefunPlugin extends JavaPlugin implements SlimefunAddon {
 
     private TickerTask ticker;
     private SlimefunCommand command;
-    private RecipeSnapshot recipeSnapshot;
 
     private Config researches;
     private Config items;
@@ -230,7 +230,7 @@ public final class SlimefunPlugin extends JavaPlugin implements SlimefunAddon {
             ancientAltarListener = new AncientAltarListener();
             grapplingHookListener = new GrapplingHookListener();
 
-            // Toggleable Listeners for performance
+            // Toggleable Listeners for performance reasons
             if (config.getBoolean("items.talismans")) {
                 new TalismanListener(this);
             }
@@ -256,8 +256,8 @@ public final class SlimefunPlugin extends JavaPlugin implements SlimefunAddon {
             Slimefun.runSync(() -> {
                 textureService.register(registry.getAllSlimefunItems());
                 permissionsService.register(registry.getAllSlimefunItems());
+                recipeService.load();
 
-                recipeSnapshot = new RecipeSnapshot(this);
                 protections = new ProtectionManager(getServer());
 
                 PostSetup.loadItems();
@@ -382,8 +382,10 @@ public final class SlimefunPlugin extends JavaPlugin implements SlimefunAddon {
 
     @Override
     public void onDisable() {
-        // CS-CoreLib wasn't loaded, just disabling
-        if (instance == null) return;
+        // Slimefun never loaded successfully, so we don't even bother doing stuff here
+        if (instance == null) {
+            return;
+        }
 
         Bukkit.getScheduler().cancelTasks(this);
 
@@ -497,8 +499,8 @@ public final class SlimefunPlugin extends JavaPlugin implements SlimefunAddon {
         return instance.local;
     }
 
-    public static RecipeSnapshot getMinecraftRecipes() {
-        return instance.recipeSnapshot;
+    public static MinecraftRecipeService getMinecraftRecipes() {
+        return instance.recipeService;
     }
 
     public static CustomItemDataService getItemDataService() {